# contracts/openapi/cellarbridge-api.yaml
openapi: 3.1.0
info:
  title: CellarBridge API
  version: 1.10.0
  summary: Contract baseline for B2B wine quotation, order, inventory, fulfillment and exception recovery
  description: |
    This contract remains the design-first integration baseline for CellarBridge. All examples are synthetic.
    The authenticated `/me`, partner-onboarding, catalog supply-search, quotation and trade-order
    endpoints, Inventory Reservation operations, Fulfillment Plan operations, and Exception Center recovery are delivered as vertical slices; remaining business endpoints retain their approved contracts.
    Internal commercial fields and customer-portal fields intentionally use separate schemas.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /exceptions:
    get:
      tags: [Exceptions]
      operationId: listExceptions
      summary: List exception cases
      x-permissions: [exception:read]
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/ExceptionStatus' }
          explode: false
        - name: severity
          in: query
          schema: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        - name: assigneeId
          in: query
          schema: { type: string, format: uuid }
        - name: sourceType
          in: query
          schema: { type: string, pattern: '^[A-Z][A-Z0-9_]{1,79}$' }
        - name: overdue
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Exception page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExceptionPage' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
  /exceptions/{exceptionId}:
    get:
      tags: [Exceptions]
      operationId: getException
      summary: Get exception detail and recovery history
      x-permissions: [exception:read]
      parameters:
        - $ref: '#/components/parameters/ExceptionId'
      responses:
        '200':
          description: Exception detail
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExceptionDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
  /exceptions/{exceptionId}/assignment:
    post:
      tags: [Exceptions]
      operationId: assignException
      summary: Assign an open exception case
      x-requirements: [FR-EXC-002]
      x-permissions: [exception:assign]
      parameters:
        - $ref: '#/components/parameters/ExceptionId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExceptionAssignmentRequest' }
      responses:
        '200':
          description: Assigned exception detail
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExceptionDetail' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /exceptions/{exceptionId}/actions:
    post:
      tags: [Exceptions]
      operationId: transitionException
      summary: Acknowledge an exception or begin its investigation
      x-requirements: [FR-EXC-002]
      x-permissions: [exception:recover]
      parameters:
        - $ref: '#/components/parameters/ExceptionId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExceptionLifecycleRequest' }
      responses:
        '200':
          description: Updated exception detail
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExceptionDetail' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /exceptions/{exceptionId}/recovery-attempts:
    post:
      tags: [Exceptions]
      operationId: requestExceptionRecovery
      summary: Execute an allowed, auditable recovery action
      x-requirements: [FR-EXC-003]
      x-permissions: [exception:recover]
      parameters:
        - $ref: '#/components/parameters/ExceptionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecoveryRequest' }
      responses:
        '202':
          description: Recovery attempt accepted
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Idempotency-Replayed: { $ref: '#/components/headers/IdempotencyReplayed' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecoveryResult' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /exceptions/{exceptionId}/closure:
    post:
      tags: [Exceptions]
      operationId: closeException
      summary: Close a verified, false-positive or duplicate exception
      x-requirements: [FR-EXC-004]
      x-permissions: [exception:recover]
      parameters:
        - $ref: '#/components/parameters/ExceptionId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExceptionClosureRequest' }
      responses:
        '200':
          description: Closed exception detail
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExceptionDetail' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /event-publications/failed:
    get:
      tags: [Exceptions]
      operationId: listFailedEventPublications
      summary: List masked failed local deliveries for controlled replay
      x-requirements: [FR-EXC-004]
      x-permissions: [event-publication:read]
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Failed delivery page without commercial payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FailedEventDeliveryPage' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    ETag:
      description: Quoted optimistic version value
      schema: { type: string, example: '"7"' }
    IdempotencyReplayed:
      description: True when an earlier completed result is returned
      schema: { type: boolean }
  parameters:
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 1000 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string, pattern: '^"[0-9]+"$' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 20
        maxLength: 200
        pattern: '^[A-Za-z0-9._~-]+$'
...
    ExceptionId:
      name: exceptionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    AuthenticationRequired:
      description: Authentication required or token invalid
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    AccessDenied:
      description: Authenticated actor is not allowed to perform the action
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource is not visible in the current tenant and permission scope
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationFailed:
      description: Request validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Business state, uniqueness or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    VersionConflict:
      description: Expected version does not match the current resource version
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionRequired:
      description: If-Match is required
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code, traceId, retryable]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
        traceId: { type: string }
        retryable: { type: boolean }
        currentVersion: { type: integer, format: int64 }
        currentState: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    ExceptionStatus:
      type: string
      enum: [OPEN, ASSIGNED, ACKNOWLEDGED, IN_PROGRESS, RECOVERY_PENDING, RESOLVED, CLOSED]
    ExceptionCategory:
      type: string
      enum: [INVENTORY_SHORTAGE, FULFILLMENT_STEP_FAILED, FULFILLMENT_STEP_OVERDUE, EVENT_DELIVERY_FAILED]
    ExceptionSeverity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    ExceptionSummary:
      type: object
      required: [id, number, category, severity, status, sourceType, sourceId, sourceNumber, summary, openedAt, version]
      properties:
        id: { type: string, format: uuid }
        number: { type: string }
        category: { $ref: '#/components/schemas/ExceptionCategory' }
        severity: { $ref: '#/components/schemas/ExceptionSeverity' }
        status: { $ref: '#/components/schemas/ExceptionStatus' }
        sourceType: { type: string }
        sourceId: { type: string, format: uuid }
        sourceNumber: { type: string }
        summary: { type: string }
        assigneeId: { type: [string, 'null'], format: uuid }
        primaryCaseId: { type: [string, 'null'], format: uuid }
        openedAt: { type: string, format: date-time }
        dueAt: { type: [string, 'null'], format: date-time }
        version: { type: integer, format: int64 }
    ExceptionPage:
      type: object
      required: [items, hasNext, pageSize]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ExceptionSummary' }
        nextCursor: { type: [string, 'null'] }
        hasNext: { type: boolean }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
    ExceptionDetail:
      type: object
      required: [summary, safeDetails, occurrences, history, recoveries, allowedRecoveryActions, allowedActions]
      properties:
        summary: { $ref: '#/components/schemas/ExceptionSummary' }
        safeDetails: { type: object, additionalProperties: true }
        occurrences:
          type: array
          items: { $ref: '#/components/schemas/ExceptionOccurrence' }
        history:
          type: array
          items: { $ref: '#/components/schemas/ExceptionHistory' }
        recoveries:
          type: array
          items: { $ref: '#/components/schemas/RecoveryAttempt' }
        allowedRecoveryActions:
          type: array
          items: { $ref: '#/components/schemas/RecoveryAction' }
        allowedActions:
          type: array
          items: { type: string }
    ExceptionOccurrence:
      type: object
      required: [sourceEventId, eventType, detectedAt, evidence]
      properties:
        sourceEventId: { type: string, format: uuid }
        eventType: { type: string }
        detectedAt: { type: string, format: date-time }
        evidence: { type: object, additionalProperties: true }
    ExceptionHistory:
      type: object
      required: [action, actorType, newStatus, correlationId, occurredAt]
      properties:
        action: { type: string }
        actorType: { type: string, enum: [SYSTEM, INTERNAL_USER] }
        actorId: { type: [string, 'null'], format: uuid }
        previousStatus:
          oneOf:
            - $ref: '#/components/schemas/ExceptionStatus'
            - type: 'null'
        newStatus: { $ref: '#/components/schemas/ExceptionStatus' }
        reasonCode: { type: [string, 'null'] }
        safeReason: { type: [string, 'null'] }
        correlationId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time }
    RecoveryAction:
      type: string
      enum: [RETRY_RESERVATION, RETRY_FULFILLMENT_STEP, RESUME_FULFILLMENT_PLAN, REPLAY_PUBLICATION, MANUAL_ACKNOWLEDGE]
    RecoveryAttempt:
      type: object
      required: [id, action, requesterId, inputSummary, requestedAt]
      properties:
        id: { type: string, format: uuid }
        action: { $ref: '#/components/schemas/RecoveryAction' }
        requesterId: { type: string, format: uuid }
        inputSummary: { type: object, additionalProperties: true }
        requestedAt: { type: string, format: date-time }
        outcome:
          oneOf:
            - $ref: '#/components/schemas/RecoveryOutcome'
            - type: 'null'
    RecoveryOutcome:
      type: object
      required: [status, resultCode, safeResult, completedAt]
      properties:
        status: { type: string, enum: [SUCCEEDED, FAILED] }
        resultCode: { type: string }
        safeResult: { type: string }
        sourceState: { type: [string, 'null'] }
        completedAt: { type: string, format: date-time }
    ExceptionAssignmentRequest:
      type: object
      required: [assigneeId, reason]
      properties:
        assigneeId: { type: string, format: uuid }
        reason: { type: string, minLength: 5, maxLength: 500 }
    ExceptionLifecycleRequest:
      type: object
      required: [action, reason]
      properties:
        action: { type: string, enum: [ACKNOWLEDGE, BEGIN_INVESTIGATION] }
        reason: { type: string, minLength: 5, maxLength: 500 }
    RecoveryRequest:
      type: object
      required: [action, reason]
      properties:
        action: { $ref: '#/components/schemas/RecoveryAction' }
        reason: { type: string, minLength: 5, maxLength: 500 }
        parameters: { type: object, additionalProperties: true }
    RecoveryResult:
      type: object
      required: [exceptionId, attemptId, status, resultCode, safeResult, caseStatus, version, completedAt, replayed]
      properties:
        exceptionId: { type: string, format: uuid }
        attemptId: { type: string, format: uuid }
        status: { type: string, enum: [SUCCEEDED, FAILED] }
        resultCode: { type: string }
        safeResult: { type: string }
        sourceState: { type: [string, 'null'] }
        caseStatus: { $ref: '#/components/schemas/ExceptionStatus' }
        version: { type: integer, format: int64 }
        completedAt: { type: string, format: date-time }
        replayed: { type: boolean }
    ExceptionClosureRequest:
      type: object
      required: [reasonCode, reason]
      properties:
        reasonCode:
          type: string
          enum: [RECOVERY_VERIFIED, FALSE_POSITIVE, DUPLICATE]
        reason: { type: string, minLength: 5, maxLength: 500 }
        primaryCaseId: { type: [string, 'null'], format: uuid }
    FailedEventDelivery:
      type: object
      required: [eventId, eventType, consumerName, status, attempts, errorCode, lastAttemptAt, version]
      properties:
        eventId: { type: string, format: uuid }
        eventType: { type: string }
        consumerName: { type: string }
        status: { type: string, enum: [FAILED_RETRYABLE, FAILED_FINAL] }
        attempts: { type: integer, minimum: 1 }
        nextRetryAt: { type: [string, 'null'], format: date-time }
        errorCode: { type: string }
        lastAttemptAt: { type: string, format: date-time }
        version: { type: integer, format: int64 }
    FailedEventDeliveryPage:
      type: object
      required: [items, hasNext, pageSize]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FailedEventDelivery' }
        nextCursor: { type: [string, 'null'] }
        hasNext: { type: boolean }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
...